pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "energy-weather-analysis"
version = "0.1.0"
description = "US Weather and Energy Analysis Pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "folium>=0.14.0",
    "streamlit-folium>=0.15.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "jupyterlab>=4.0.0",
]

[project.scripts]
fetch-data = "src.pipeline:main"
run-dashboard = "dashboards.app:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true